Consolidate Connection, SynConnection, and RedisServer.
