+++ plugins/redis/lib_redis.ecllib
@@ -44,7 +44,7 @@ EXPORT sync := SERVICE : plugin('redis'), namespace('RedisPlugin')
   INTEGER DBSize(CONST VARSTRING options, UNSIGNED database = 0, CONST VARSTRING password = '', UNSIGNED timeout = 1000000) : cpp,once,context,entrypoint='RDBSize';
 END;
 
-EXPORT syncLockingInternal := SERVICE : plugin('redis'), namespace('RedisPlugin')
+EXPORT locking := SERVICE : plugin('redis'), namespace('RedisPlugin')
   UNICODE  SetUnicode( CONST VARSTRING key, CONST UNICODE value, CONST VARSTRING options, UNSIGNED database = 0, UNSIGNED4 expire = 0, CONST VARSTRING password = '', UNSIGNED timeout = 1000000) : cpp,once,context,entrypoint='SyncLockRSetUChar';
   STRING   SetString(  CONST VARSTRING key, CONST STRING value,  CONST VARSTRING options, UNSIGNED database = 0, UNSIGNED4 expire = 0, CONST VARSTRING password = '', UNSIGNED timeout = 1000000) : cpp,once,context,entrypoint='SyncLockRSetStr';
   UTF8     SetUtf8(    CONST VARSTRING key, CONST UTF8 value,    CONST VARSTRING options, UNSIGNED database = 0, UNSIGNED4 expire = 0, CONST VARSTRING password = '', UNSIGNED timeout = 1000000) : cpp,once,context,entrypoint='SyncLockRSetUtf8';
@@ -96,36 +96,25 @@ EXPORT RedisServer    (VARSTRING options, VARSTRING password = '', UNSIGNED time
 
   //locking
   EXPORT locking := MODULE
-    EXPORT  SetUnicode(VARSTRING key, UNICODE value,  UNSIGNED database = 0, UNSIGNED4 expire = 0) := syncLockingInternal.SetUnicode (key, value, options, database, expire, password, timeout);
-    EXPORT   SetString(VARSTRING key, STRING value,   UNSIGNED database = 0, UNSIGNED4 expire = 0) := syncLockingInternal.SetString  (key, value, options, database, expire, password, timeout);
-    EXPORT     SetUtf8(VARSTRING key, UTF8 value,     UNSIGNED database = 0, UNSIGNED4 expire = 0) := syncLockingInternal.SetUtf8    (key, value, options, database, expire, password, timeout);
-    EXPORT  SetBoolean(VARSTRING key, BOOLEAN value,  UNSIGNED database = 0, UNSIGNED4 expire = 0) := syncLockingInternal.SetBoolean (key, value, options, database, expire, password, timeout);
-    EXPORT     SetReal(VARSTRING key, REAL value,     UNSIGNED database = 0, UNSIGNED4 expire = 0) := syncLockingInternal.SetReal    (key, value, options, database, expire, password, timeout);
-    EXPORT  SetInteger(VARSTRING key, INTEGER value,  UNSIGNED database = 0, UNSIGNED4 expire = 0) := syncLockingInternal.SetInteger (key, value, options, database, expire, password, timeout);
-    EXPORT SetUnsigned(VARSTRING key, UNSIGNED value, UNSIGNED database = 0, UNSIGNED4 expire = 0) := syncLockingInternal.SetUnsigned(key, value, options, database, expire, password, timeout);
-    EXPORT     SetData(VARSTRING key, DATA value,     UNSIGNED database = 0, UNSIGNED4 expire = 0) := syncLockingInternal.SetData    (key, value, options, database, expire, password, timeout);
+    EXPORT  SetUnicode(VARSTRING key, UNICODE value,  UNSIGNED database = 0, UNSIGNED4 expire = 0) := locking.SetUnicode (key, value, options, database, expire, password, timeout);
+    EXPORT   SetString(VARSTRING key, STRING value,   UNSIGNED database = 0, UNSIGNED4 expire = 0) := locking.SetString  (key, value, options, database, expire, password, timeout);
+    EXPORT     SetUtf8(VARSTRING key, UTF8 value,     UNSIGNED database = 0, UNSIGNED4 expire = 0) := locking.SetUtf8    (key, value, options, database, expire, password, timeout);
+    EXPORT  SetBoolean(VARSTRING key, BOOLEAN value,  UNSIGNED database = 0, UNSIGNED4 expire = 0) := locking.SetBoolean (key, value, options, database, expire, password, timeout);
+    EXPORT     SetReal(VARSTRING key, REAL value,     UNSIGNED database = 0, UNSIGNED4 expire = 0) := locking.SetReal    (key, value, options, database, expire, password, timeout);
+    EXPORT  SetInteger(VARSTRING key, INTEGER value,  UNSIGNED database = 0, UNSIGNED4 expire = 0) := locking.SetInteger (key, value, options, database, expire, password, timeout);
+    EXPORT SetUnsigned(VARSTRING key, UNSIGNED value, UNSIGNED database = 0, UNSIGNED4 expire = 0) := locking.SetUnsigned(key, value, options, database, expire, password, timeout);
+    EXPORT     SetData(VARSTRING key, DATA value,     UNSIGNED database = 0, UNSIGNED4 expire = 0) := locking.SetData    (key, value, options, database, expire, password, timeout);
 
-    EXPORT  GetUnicode(VARSTRING key, UNSIGNED database = 0) :=  syncLockingInternal.GetUnicode(key, options, database, password, timeout);
-    EXPORT   GetString(VARSTRING key, UNSIGNED database = 0) :=   syncLockingInternal.GetString(key, options, database, password, timeout);
-    EXPORT     GetUtf8(VARSTRING key, UNSIGNED database = 0) :=     syncLockingInternal.GetUtf8(key, options, database, password, timeout);
-    EXPORT  GetBoolean(VARSTRING key, UNSIGNED database = 0) :=  syncLockingInternal.GetBoolean(key, options, database, password, timeout);
-    EXPORT     GetReal(VARSTRING key, UNSIGNED database = 0) :=     syncLockingInternal.GetReal(key, options, database, password, timeout);
-    EXPORT  GetInteger(VARSTRING key, UNSIGNED database = 0) :=  syncLockingInternal.GetInteger(key, options, database, password, timeout);
-    EXPORT GetUnsigned(VARSTRING key, UNSIGNED database = 0) := syncLockingInternal.GetUnsigned(key, options, database, password, timeout);
-    EXPORT     GetData(VARSTRING key, UNSIGNED database = 0) :=     syncLockingInternal.GetData(key, options, database, password, timeout);
+    EXPORT  GetUnicode(VARSTRING key, UNSIGNED database = 0) :=  locking.GetUnicode(key, options, database, password, timeout);
+    EXPORT   GetString(VARSTRING key, UNSIGNED database = 0) :=   locking.GetString(key, options, database, password, timeout);
+    EXPORT     GetUtf8(VARSTRING key, UNSIGNED database = 0) :=     locking.GetUtf8(key, options, database, password, timeout);
+    EXPORT  GetBoolean(VARSTRING key, UNSIGNED database = 0) :=  locking.GetBoolean(key, options, database, password, timeout);
+    EXPORT     GetReal(VARSTRING key, UNSIGNED database = 0) :=     locking.GetReal(key, options, database, password, timeout);
+    EXPORT  GetInteger(VARSTRING key, UNSIGNED database = 0) :=  locking.GetInteger(key, options, database, password, timeout);
+    EXPORT GetUnsigned(VARSTRING key, UNSIGNED database = 0) := locking.GetUnsigned(key, options, database, password, timeout);
+    EXPORT     GetData(VARSTRING key, UNSIGNED database = 0) :=     locking.GetData(key, options, database, password, timeout);
 
-    EXPORT Exists(VARSTRING key, UNSIGNED database = 0) := NOT(syncLockingInternal.MissTHenLock(key, options, database, password, timeout));
-    EXPORT Unlock(VARSTRING key, UNSIGNED database = 0) := syncLockingInternal.unlock(key, options, database, password, timeout);
-
-
-//----------------WIP--------------------------------
-/*
-    EXPORT GetString2(key, evalFunc, database, expire) := FUNCTIONMACRO
-        IF (syncLockingInternal.MissThenLock(key, options, database, password, timeout),
-            syncLockingInternal.SetString(key, evalFunc, options, database, expire, password, timeout),
-            syncLockingInternal.GetString(key, options, database, password, timeout)
-        );
-    ENDMACRO;
-*/
+    EXPORT Exists(VARSTRING key, UNSIGNED database = 0) := NOT(locking.MissTHenLock(key, options, database, password, timeout));
+    EXPORT Unlock(VARSTRING key, UNSIGNED database = 0) := locking.unlock(key, options, database, password, timeout);
    END;
 END;
